make installation of requirements based on test, introduce env variable for plugin name (PLUGIN)

+++ bin/travis-build.bash
@@ -1,13 +1,13 @@
 #!/bin/bash
 set -e
 
-echo "This is travis-build.bash..."
+echo "This is travis-build.bash ..."
 
-echo "Installing the packages that CKAN requires..."
+echo "Installing the packages that CKAN requires ..."
 sudo apt-get update -qq
 sudo apt-get install solr-jetty
 
-echo "Installing CKAN and its Python dependencies..."
+echo "Cloning CKAN source ($CKANVERSION) from repository ..."
 git clone https://github.com/ckan/ckan
 cd ckan
 if [ $CKANVERSION == 'master' ]
@@ -22,26 +22,27 @@ fi
 # install the recommended version of setuptools
 if [ -f requirement-setuptools.txt ]
 then
-    echo "Updating setuptools..."
+    echo "Updating setuptools ..."
     pip install -r requirement-setuptools.txt
 fi
 
 if [ $CKANVERSION == '2.7' ]
 then
-    echo "Installing setuptools"
+    echo "Installing setuptools ..."
     pip install setuptools==39.0.1
 fi
 
+echo "Installing CKAN and its Python dependencies ..."
 python setup.py develop
 pip install -r requirements.txt
 pip install -r dev-requirements.txt
 cd -
 
-echo "Creating the PostgreSQL user and database..."
+echo "Creating the PostgreSQL user and database ..."
 sudo -u postgres psql -c "CREATE USER ckan_default WITH PASSWORD 'pass';"
 sudo -u postgres psql -c 'CREATE DATABASE ckan_test WITH OWNER ckan_default;'
 
-echo "SOLR config..."
+echo "SOLR config ..."
 # Solr is multicore for tests on ckan master, but it's easier to run tests on
 # Travis single-core. See https://github.com/ckan/ckan/issues/2972
 sed -i -e 's/solr_url.*/solr_url = http:\/\/127.0.0.1:8983\/solr/' ckan/test-core.ini
@@ -49,14 +50,21 @@ printf "NO_START=0\nJETTY_HOST=127.0.0.1\nJETTY_PORT=8983\nJAVA_HOME=$JAVA_HOME"
 sudo cp ckan/ckan/config/solr/schema.xml /etc/solr/conf/schema.xml
 sudo service jetty restart
 
-echo "Initialising the database..."
+echo "Initialising the database ..."
 cd ckan
 paster db init -c test-core.ini
 cd -
 
-echo "Installing ckanext-fisbroker and its requirements..."
+echo "Installing $PLUGIN and its requirements ..."
 python setup.py develop
-pip install -r dev-requirements.txt
+if [ -f requirements.txt ]
+then
+    pip install -r requirements.txt
+fi
+if [ -f dev-requirements.txt ]
+then
+    pip install -r dev-requirements.txt
+fi
 
 echo "Moving test.ini into a subdir..."
 mkdir subdir
